+++ src/renderer/src/atoms/window.ts
@@ -1,0 +1,11 @@
+import { atom, useAtomValue } from 'jotai'
+
+export enum WindowState {
+  MINIMIZED = 'minimized',
+  MAXIMIZED = 'maximized',
+  NORMAL = 'normal',
+}
+
+export const windowStateAtom = atom<WindowState>(WindowState.NORMAL)
+
+export const useWindowState = () => useAtomValue(windowStateAtom)