+++ src/renderer/src/components/modules/settings/views/general/General.tsx
@@ -29,6 +29,16 @@ export const GeneralView = () => {
           <DarkModeToggle />
         </FieldLayout>
 
+        {!isWeb && (
+          <FieldLayout title="播放记录显示海报" className="pt-1">
+            <SettingSwitch
+              value={appSettings.showPoster}
+              onCheckedChange={(checked) =>
+                setAppSettings((prev) => ({ ...prev, showPoster: checked }))
+              }
+            />
+          </FieldLayout>
+        )}
       </FieldsCardLayout>
     </SettingViewContainer>
   )