+++ src/renderer/src/components/modules/settings/views/player/DanmakuSetting.tsx
@@ -1,5 +1,6 @@
 import { usePlayerSettings } from '@renderer/atoms/settings/player'
 import { SettingSelect } from '@renderer/components/modules/shared/setting/SettingSelect'
+import { SettingSwitch } from '@renderer/components/modules/shared/setting/SettingSwitch'
 import type { FC, PropsWithChildren } from 'react'
 import { useMemo } from 'react'
 
@@ -12,29 +13,31 @@ interface DanmakuSettingProps extends PropsWithChildren {
 }
 
 export const DanmakuSetting: FC<DanmakuSettingProps> = (props) => {
-  const { classNames, children } = props
+  const { classNames, children, onTraditionalToSimplifiedChange } = props
   const [playerSetting, setPlayerSetting] = usePlayerSettings()
+  const isPlaying = !!classNames?.cardLayout
   const CardLayout = useMemo(() => {
     return ({ children }) =>
-      classNames?.cardLayout ? (
+      isPlaying ? (
         <div className={classNames?.cardLayout}>{children}</div>
       ) : (
         <FieldsCardLayout title="弹幕">{children}</FieldsCardLayout>
       )
-  }, [classNames])
+  }, [classNames?.cardLayout, isPlaying])
 
-  // TODO 完成繁体转简体
   return (
     <CardLayout>
-      {/* <FieldLayout title="繁体转简体">
-        <SettingSwitch
-          value={playerSetting.enableTraditionalToSimplified}
-          onCheckedChange={(value) => {
-            setPlayerSetting((prev) => ({ ...prev, enableTraditionalToSimplified: value }))
-            onTraditionalToSimplifiedChange?.(value)
-          }}
-        />
-      </FieldLayout> */}
+      {!isPlaying && (
+        <FieldLayout title="繁体转简体">
+          <SettingSwitch
+            value={playerSetting.enableTraditionalToSimplified}
+            onCheckedChange={(value) => {
+              setPlayerSetting((prev) => ({ ...prev, enableTraditionalToSimplified: value }))
+              onTraditionalToSimplifiedChange?.(value)
+            }}
+          />
+        </FieldLayout>
+      )}
       <FieldLayout title="字体大小">
         <SettingSelect
           placeholder="弹幕字体大小"