+++ src/renderer/src/components/ui/toast/use-toast.tsx
@@ -190,4 +190,4 @@ function useToast() {
   }
 }
 
-export { toast,useToast }
+export { toast, useToast }