+++ src/renderer/src/database/db.ts
@@ -27,6 +27,15 @@ class LocalDB extends Dexie {
     this.history = this.table(TABLES.HISTORY)
     this.bangumi = this.table(TABLES.BANGUMI)
   }
+
+  async deleteDatabase(): Promise<void> {
+    try {
+      db.close()
+      await Dexie.delete(db.name)
+    } catch (error) {
+      console.error('删除数据库失败:', error)
+    }
+  }
 }
 
 export const db = new LocalDB()