+++ src/renderer/src/initialize/sentry.ts
@@ -1,5 +1,8 @@
+import { version } from '@pkg'
 import { API_URL, isDev, SENTRY_DSN } from '@renderer/lib/env'
 import * as Sentry from '@sentry/react'
+import { useEffect } from 'react'
+import { createRoutesFromChildren, matchRoutes, useLocation, useNavigationType } from 'react-router'
 
 export const initializeSentry = () => {
   if (isDev) {
@@ -7,7 +10,24 @@ export const initializeSentry = () => {
   }
   Sentry.init({
     dsn: SENTRY_DSN,
-    integrations: [Sentry.browserTracingIntegration(), Sentry.replayIntegration()],
+    integrations: [
+      Sentry.reactRouterV7BrowserTracingIntegration({
+        useEffect,
+        useLocation,
+        useNavigationType,
+        createRoutesFromChildren,
+        matchRoutes,
+      }),
+      Sentry.replayIntegration({
+        // NOTE: This will disable built-in masking. Only use this if your site has no sensitive data, or if you've already set up other options for masking or blocking relevant data, such as 'ignore', 'block', 'mask' and 'maskFn'.
+        maskAllText: false,
+        blockAllMedia: false,
+      }),
+      Sentry.httpClientIntegration(),
+      Sentry.captureConsoleIntegration({
+        levels: ['error'],
+      }),
+    ],
     // Tracing
     tracesSampleRate: 1, //  Capture 100% of the transactions
     // Set 'tracePropagationTargets' to control for which URLs distributed tracing should be enabled
@@ -16,4 +36,5 @@ export const initializeSentry = () => {
     replaysSessionSampleRate: 0.1, // This sets the sample rate at 10%. You may want to change it to 100% while in development and then sample at a lower rate in production.
     replaysOnErrorSampleRate: 1, // If you're not already sampling the entire session, change the sample rate to 100% when sampling sessions where errors occur.
   })
+  Sentry.setTag('app_version', version)
 }